Move Interface declaration to a distinct header file
