+++ include/vrv/pitchinterface.h
@@ -10,6 +10,7 @@
 
 #include "atts_gestural.h"
 #include "atts_shared.h"
+#include "interface.h"
 
 namespace vrv {
