+++ src/options.cpp
@@ -19,6 +19,7 @@
 
 #include "att.h"
 #include "vrv.h"
+#include "vrvdef.h"
 
 namespace vrv {
